Update finance calculator

+++ python_introduction/finance_calculator.py
@@ -1,8 +1,8 @@
-user_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+user_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 monthly_savings = user_income - total_monthly_expenses
-projected_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+projected_savings = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print(f"Your monthly savings are ${monthly_savings}.")
